Extract math help modal into extra file and use ModalWithCloseButton
