+++ src/serlo-editor/math/math-help-modal.tsx
@@ -1,5 +1,4 @@
-import Modal from 'react-modal'
-
+import { ModalWithCloseButton } from '@/components/modal-with-close-button'
 import { useEditorStrings } from '@/contexts/logged-in-data-context'
 
 export const MathHelpModal = ({
@@ -11,31 +10,9 @@ export const MathHelpModal = ({
 }) => {
   const mathStrings = useEditorStrings().plugins.text.math
   return (
-    <Modal
-      ariaHideApp={false}
+    <ModalWithCloseButton
       isOpen={isHelpOpen}
-      onRequestClose={() => void setIsHelpOpen(false)}
-      style={{
-        overlay: {
-          zIndex: 9999,
-          display: 'flex',
-          justifyContent: 'center',
-          alignItems: 'center',
-          backgroundColor: 'rgba(0, 0, 0, 0.1)',
-        },
-        content: {
-          borderRadius: 0,
-          backgroundColor: '#ffffff',
-          width: '90%',
-          maxWidth: '600px',
-          maxHeight: 'calc(100vh - 80px)',
-          top: 'auto',
-          left: 'auto',
-          right: 'auto',
-          bottom: 'auto',
-          margin: '0 auto',
-        },
-      }}
+      onCloseClick={() => void setIsHelpOpen(false)}
     >
       <>
         {mathStrings.shortcuts}:
@@ -71,7 +48,7 @@ export const MathHelpModal = ({
           {renderKey('ln')}, …
         </p>
       </>
-    </Modal>
+    </ModalWithCloseButton>
   )
 
   function renderKey(text: string) {
